pyproject.toml
[project]
name = "django-mozza"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.16",
    "django-debug-toolbar>=6.0.0",
    "djlint>=1.36.4",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.5",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "website.settings"
python_files = ["*_test.py"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]